+++ Source/BlenderEditorControlsPlugin/Public/Tools/MoveTool.h
@@ -12,7 +12,7 @@ namespace BlenderControls
 
         /* FBlenderToolBase */
         virtual void OnActive(const FVector2D &CurrentViewportMousePosition) override;
-        virtual void ApplyNumeric(float Value) override;
+        virtual void ApplyNumeric(double Value) override;
         virtual void UpdateHud() override;
 
         virtual void OnBegin() override;
