+++ Source/BlenderEditorControlsPlugin/Public/Tools/ScaleTool.h
@@ -15,7 +15,7 @@ namespace BlenderControls
 
 		/* ---------- FBlenderToolBase overrides ---------- */
 		virtual void OnActive(const FVector2D &CurrentViewportMousePosition) override;
-		virtual void ApplyNumeric(float Value) override;
+		virtual void ApplyNumeric(double Value) override;
 		virtual void UpdateHud() override;
 		virtual void OnBegin() override;
 		virtual void OnEnd(bool bApply) override;
